Inject the group id in the consumer factory
Since the group id decides whether a consumer will be able to subscribe
to a partition within a topic, and whether or not that consumer will
actually get a certain message, the group id needs to be configurable.

For example:
If 2 developers would use the exact same group id, they would not be
able to subscribe to the same topics and partitions. That would cause
one of them to get one half of the messages, while the other would get
the other half. This causes a buggy experience as only half of the
expected functionality works, or only half the time.