Require `--topic=<topic>` or `--all-topics` to start consumer
The previous behaviour of the consumer would look at the environment name to switch between subscribing to all topics vs subscribing to KAFKA_TOPIC.

With this change, we force the consumer to be either executed with `--topic=<topic>` or `--all-topics`.

The later is more dangerous and should only be run in dev and test environments.